Misc Fixes
Fixed single activity display axis labels and rounding issues
Added short delay when switching between chart types
Fixed the case when a resample PDF is processed when the orginial PDF was not processed properly, when this occurs only the resample results are insertted, this also allows the water quality map to grab older records when newer records are missing.
Tried to fix logger debug error in which the logger fails to handle the success md5 insertion message.