Made tracker dashboard timezone aware and removed SSL
Set livetracker to be timezone aware, accepts timezone string in POST message, inserts it into Postgres, and uses it with the tracker API and website to determine what to return for gps tracks.

Removed LetsEncrypt certbot from project, would install when application was deployed, but would fail to auto-refresh certification. I believe this usage of certbot has been depreciated. Solutions from the GIST failed to work properly, will setup CloudFlare for now and revisit later to see if a valid solution is available.