OSM Overpass API and fixes to multithreading
Liveviewer now uses the OSM Overpass API in addition to GeoAlchemy to determine location, and will fail over to GeoAlchemy. This should allow the viewer to grab a location anywhere in the world.

Working on making the liveviewer page show movement tracks for the day in the recorded timezone, instead of UTC. Tracker API now takes the local timezone as a argument in the GPSLogger POST request.

Trying to get multithreading working to speed up response to Strava POST requests. Flask now starts a thread to handle processing the newly added activity after the POST request has been verified and the update has been commited to Postgres. The HTTP request context isn't needed anymore since the new activity ID can be pulled out seperately and passed into the thread.