Refactoring util to be packages
Refactored folders in util dir to be packages, by doing so pygeoapi can properly generate the openapi.yml file. The tool could not handle references to "application" so this is a fix, consider seeing if I can force it to reference "application" in imports.

Moved ortho extentions to util folder.

Added helper .bat files to assist in generating config/settings files